🚚  : 入力フォームの共通化
